+++ Client/src/react-app-env.d.ts
@@ -1,1 +1,2 @@
 /// <reference types="react-scripts" />
+declare module '*.png'
